+++ asteroid-core/src/main/java/asteroid/global/ClassNodeTransformer.java
@@ -1,9 +1,12 @@
 package asteroid.global;
 
-import org.codehaus.groovy.ast.expr.Expression;
+import static org.codehaus.groovy.runtime.DefaultGroovyMethods.any;
+
 import org.codehaus.groovy.ast.ClassNode;
+import org.codehaus.groovy.ast.AnnotationNode;
 
 import org.codehaus.groovy.control.SourceUnit;
+import org.codehaus.groovy.control.CompilePhase;
 import groovy.lang.Closure;
 
 /**
@@ -26,7 +29,7 @@ import groovy.lang.Closure;
  */
 public abstract class ClassNodeTransformer extends Transformer {
 
-    private Closure<Boolean> criteria;
+    private final Closure<Boolean> criteria;
 
     /**
      * Every instance needs the source unit and the name of the class
@@ -50,7 +53,7 @@ public abstract class ClassNodeTransformer extends Transformer {
      * @since 0.1.2
      * @see ClassNodeTransformer#byNameContains
      */
-    public ClassNodeTransformer(SourceUnit sourceUnit, Closure<Boolean> criteria) {
+    public ClassNodeTransformer(final SourceUnit sourceUnit, final Closure<Boolean> criteria) {
         super(sourceUnit);
         this.criteria = criteria;
     }
@@ -65,7 +68,7 @@ public abstract class ClassNodeTransformer extends Transformer {
      */
     public static Closure<Boolean> byNameContains(final String term) {
         return new Closure<Boolean>(null) {
-            public Boolean doCall(ClassNode node) {
+            public Boolean doCall(final ClassNode node) {
                 return node.getName().contains(term);
             }
         };
@@ -81,7 +84,7 @@ public abstract class ClassNodeTransformer extends Transformer {
      */
     public static Closure<Boolean> byNameEndsWith(final String term) {
         return new Closure<Boolean>(null) {
-            public Boolean doCall(ClassNode node) {
+            public Boolean doCall(final ClassNode node) {
                 return node.getName().endsWith(term);
             }
         };
@@ -97,8 +100,37 @@ public abstract class ClassNodeTransformer extends Transformer {
      */
     public static Closure<Boolean> byNameStartsWith(final String term) {
         return new Closure<Boolean>(null) {
-            public Boolean doCall(ClassNode node) {
+            public Boolean doCall(final ClassNode node) {
                 return node.getName().startsWith(term);
+            }
+        };
+    }
+
+    /**
+     * Criteria to find those classes with an annotation with a {@link
+     * Class} with a name as the passed argument. This name should be
+     * the same as using {@link Class#getSimpleName}
+     * <br><br>
+     * This method doesn't use a {@link Class} as argument cause the
+     * package (type information) won't be available for earlier
+     * {@link CompilePhase}
+     *
+     * @param simpleName the simple name of the {@link Class} of the annotation used as marker
+     * @return a criteria to use in the {@link ClassNodeTransformer} constructor
+     * @since 0.1.6
+     */
+    public static Closure<Boolean> byAnnotationName(final String simpleName) {
+        return new Closure<Boolean>(null) {
+            public Boolean doCall(final ClassNode node) {
+                final java.util.List<AnnotationNode> list = node.getAnnotations();
+                final boolean cond  = list != null && !list.isEmpty();
+                final boolean total = cond && any(list, new Closure(null) {
+                        public boolean doCall(final AnnotationNode node) {
+                            return node.getClassNode().getName().equals(simpleName);
+                        }
+                    });
+
+                return total;
             }
         };
     }
@@ -108,7 +140,9 @@ public abstract class ClassNodeTransformer extends Transformer {
      */
     @Override
     public void visitClass(final ClassNode classNode) {
-        if (classNode == null || !this.criteria.call(classNode)) return;
+        if (classNode == null || !this.criteria.call(classNode)) {
+            return;
+        }
 
         transformClass(classNode);
     }
@@ -120,6 +154,6 @@ public abstract class ClassNodeTransformer extends Transformer {
      * @param classNode the {@link ClassNode}  you want to transform
      * @since 0.1.2
      */
-    public abstract void transformClass(ClassNode classNode);
+    public abstract void transformClass(final ClassNode classNode);
 
 }