+++ asteroid-core/src/main/java/asteroid/utils/MethodCallExpressionUtils.java
@@ -18,12 +18,12 @@ public final class MethodCallExpressionUtils {
     /**
      * Given a {@link MethodCallExpression} it returns a list of arguments
      *
-     * @param methodCallExpression a method call we want the arguments from
+     * @param methodCallExpr a method call we want the arguments from
      * @return a list of expressions within a {@link ArgumentListExpression}
      * @since 0.1.4
      */
-    public ArgumentListExpression getArgs(final MethodCallExpression methodCallExpression) {
-        return (ArgumentListExpression) methodCallExpression.getArguments();
+    public ArgumentListExpression getArgs(final MethodCallExpression methodCallExpr) {
+        return (ArgumentListExpression) methodCallExpr.getArguments();
     }
 
     /**