+++ asteroid-test/src/test/groovy/asteroid/utils/AddImportSpecExample.groovy
@@ -1,0 +1,8 @@
+package asteroid.utils
+
+@AddImport
+class AddImportSpecExample {
+    String asJson(Map data) {
+        return JsonOutput.toJson(data)
+    }
+}